viced: calloc stats in SRXAFS_GetStatistics64
Since it was introduced in commit
b651ece9bdb2ea9dc19907a5686e9d084f3c03d6
(large-partition-support-20080305), SRXAFS_GetStatistics64 allocates
the stats memory with malloc(), and never sets the STATS64_BOOTTIME
stat. As a result, the value for STATS64_BOOTTIME contains
uninitialized heap data when sent over the wire.

To fix this, change the stats to be calloc'd instead of malloc'd.

FIXES 135444

Change-Id: Iff8ffb5255d9fbd63aa3d1d0da5e8b093e37351a
Reviewed-on: https://gerrit.openafs.org/16145
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>