fs: Still print fid for 'getfid' if GetCell fails
Currently, 'fs getfid' fails if we cannot get the cell for the given
file via GetCell(). This GetCell call was added in commit d390df097c (fs
getfid output changed for consistency with Windows implementation) to be
more similar to the WINNT 'fs getfid', but the WINNT 'fs getfid' still
prints the fid if getting the cell fails (and has since it was
introduced in commit 5520747790 (windows-fs-getfid-20090511)).

GetCell() shouldn't normally fail if getting the fid succeeded, but in
case it does, don't prevent us from printing the fid. Change 'fs
getfid' to just say the cell is "unknown-cell" and keep going.

Change-Id: Idd4727304061e1ec4eeddd98bd9eaab5de96e2b6
Reviewed-on: https://gerrit.openafs.org/15585
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>