budb: Remove superfluous rx_SetRxDeadTime
Since the original IBM code import, the buserver has set its Rx dead
time to 60 seconds at initialization.  Not only is this unnecesarily
long, but it is also moot because a subsequent call to rx_InitHost
resets the rx dead time to the default 12 seconds.

Remove the superfluous rx_SetRxDeadTime.

Change-Id: I21acaf3cef1c80af06fa031ca2cdecd29d0fc5a3
Reviewed-on: https://gerrit.openafs.org/15592
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: Benjamin Kaduk <kaduk@mit.edu>