LINUX: Refactor afs_osi_Sleep
Instead of having afs_osi_Sleep() call afs_osi_SleepSig() directly,
have both afs_osi_Sleep() and afs_osi_SleepSig() call a common helper
function, afs_linux_sleep(). This makes it easier for future commits
to alter the behavior for the _Sleep and _SleepSig variants.

This commit should incur no noticeable change in behavior. We now
manipulate the signal mask outside of AFS_GLOCK, but this doesn't matter
because 'current' and the signal mask are unrelated to any of our locks.
The signal mask is protected by SIG_LOCK (a wrapper from osi_machdep.h
for various different locks for different kernel versions), and is
handled in this commit the same as it was before.

Change-Id: Id14c44b22cf2a1883deaf6ceec66f71f3f4778a6
Reviewed-on: https://gerrit.openafs.org/15636
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>