build: clean up some more generated files
Commit c1d39153da00d5525b2f7874b2d214a7f1b1bb86 added a cc-wrapper
script to help generate CTF data for all objects, which is generated
by configure.  Remove it in the distclean target so we don't leave
a stray file hanging around.

Commit 48ce41a447c354b8a20b769e4aa5b502ba5bcc09 added a pkgbuild.sh
script, to be used to build packages for newer versions of macOS,
supplementing the buildpkg.sh that was previously used.  Clean it
up in distclean as well (buildpkg.sh was already listed).

Also clean up AFS_component_version_number.c in
libuafs/Makefile.common.in.

These issues were detected while resolving
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1048470 .

Change-Id: I0fea00924a998ee316d48cf76103ea9871a34c6b
Reviewed-on: https://gerrit.openafs.org/15605
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>