CODING: general tidying
Try to normalize the use of top-level (==) and second-level (--)
heading markers, and make the length of the header marker match
the length of the preceding line.

Wrap some long lines.

Document the existing practice of using braces for one-line conditional
bodies when other branches of the conditional need braces.

Update list of build-dependencies for Debian.

Update size of the git repository.

Mention the Stable Release Manager for backports.

Clarify that --enable-checking=all is not generally useful.

Promote the guidance to document new warning inhibitions into a
full directive (i.e., remove "please").

Add some more line breaks in the section on inhibited warnings to
try to have things line up better in the designated columns.

Change-Id: Ibfb7f3a2ad0ec82a401b4b028ab108f87a9b2b03
Reviewed-on: https://gerrit.openafs.org/15610
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>