afs: Remove DFlushDCache()
The function DFlushDcache() is not referenced anywhere within the
source tree.

Remove the function DFlushDCache().

The commit: 'disconnected-shadow-directory-fixes-20090121' (4045f3d535)
removed the code reference to DFlushDCache() and the commit:
"dir: Prototype and function name cleanup" (5ad1e6cb90) removed the
function prototype.  The function was introduced in the commit:
"disconnected-flush-before-shadowing-20090119" (e1cc987ea5).

This function is flagged due to a missing prototype when building
against a Linux 6.8 kernel (which sets the -Wmissing-declarations and
-Wmissing-prototypes compiler flags as default). Linux 6.8 commit:
 "Makefile.extrawarn: turn on missing-prototypes globally" (0fcb70851f).

When building against a kernel with CONFIG_WERROR=y, the build fails.

Change-Id: I2e8e7588c13459fe5695c68bdc031d246258e6fa
Reviewed-on: https://gerrit.openafs.org/15615
Tested-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>